add both_apis to conftest to test demo and sql apis at once
